Convert helpers to reusable UMD module
