+++ examples/aggregation_tests.html
@@ -33,7 +33,7 @@
           crossorigin="anonymous"></script>
 
   <!-- Helpers that use the external library to add functionality to LocusZoom -->
-  <script src="js/aggregation-test-helpers.js" type="application/javascript"></script>
+  <script src="../dist/ext/lz-aggregation-tests.min.js" type="application/javascript"></script>
 
   <title>LocusZoom.js ~ Aggregation Tests</title>
